+++ lib/src/phi.dart
@@ -1,5 +1,4 @@
 import 'package:control_flow_graph/control_flow_graph.dart';
-import 'package:control_flow_graph/src/ssa.dart';
 import 'package:control_flow_graph/src/types.dart';
 import 'package:more/more.dart';
 
@@ -34,43 +33,65 @@ void insertPhiNodesInto(Map<int, BasicBlock> ids, Map<String, Set<int>> globals,
 // removal of phi nodes after SSA transformation, optimizations etc
 void removePhiNodesFrom(
     CFG graph,
-    Graph<int, int> djGraph,
+    Graph<SpecifiedOperation, void> ssaGraph,
     Map<int, BasicBlock> ids,
     int root,
     Operation Function(SSA left, SSA right) assign) {
   for (var blockId in graph.depthFirst(root)) {
     final phiNodes = <PhiNode>{};
     final assignments = <int, Set<Operation>>{};
+    final replacements = <int, Map<Operation, SSA>>{};
     final block = ids[blockId]!;
     for (final op in block.code) {
       if (op is PhiNode) {
         phiNodes.add(op);
         for (final predecessor in graph.predecessorsOf(blockId)) {
           assignments[predecessor] ??= {};
-          assignments[predecessor]!.add(assign(
-              op.target,
-              findVariableInSSAGraph(
-                  ids, djGraph, predecessor, op.target.name)));
+          replacements[predecessor] ??= {};
+          final preds =
+              ssaGraph.predecessorsOf(SpecifiedOperation(blockId, op));
+          final pred =
+              preds.firstWhere((element) => element.blockId == predecessor);
+          final src = pred.op.writesTo!;
+          if (src != op.target) {
+            final ps = ssaGraph.successorsOf(pred);
+            if (ps.length == 1 && pred.op is! PhiNode) {
+              replacements[predecessor]![pred.op] = op.target;
+            } else {
+              assignments[predecessor]!.add(assign(op.target, src));
+            }
+          }
         }
       } else {
         break;
       }
     }
 
-    for (var i = 0; i < phiNodes.length; i++) {
-      block.code.removeAt(0);
-    }
-
     for (final assignment in assignments.entries) {
       final predecessor = ids[assignment.key]!;
-      final length = predecessor.code.length;
+      final code = predecessor.code;
+      final length = code.length;
       for (final op in assignment.value) {
-        if (predecessor.code[length - 1].writesTo == ControlFlowGraph.branch) {
-          predecessor.code.insert(length - 1, op);
+        if (length > 0 &&
+            code[length - 1].writesTo == ControlFlowGraph.branch) {
+          code.insert(length - 1, op);
         } else {
-          predecessor.code.add(op);
+          code.add(op);
         }
       }
+    }
+
+    for (final replacement in replacements.entries) {
+      final predecessor = ids[replacement.key]!;
+      final code = predecessor.code;
+      for (final entry in replacement.value.entries) {
+        final index = code.indexWhere((element) => element == entry.key);
+        code[index] = code[index].copyWith(writesTo: entry.value);
+      }
+    }
+
+    for (var i = 0; i < phiNodes.length; i++) {
+      block.code.removeAt(0);
     }
   }
 }