+++ test/control_flow_graph_test.dart
@@ -163,6 +163,16 @@ return x₁\n
 '''));
     });
 
+    test('Run copy propagation', () {
+      if (!cfg.hasPhiNodes) {
+        cfg.insertPhiNodes();
+      }
+      if (!cfg.inSSAForm) {
+        cfg.computeSemiPrunedSSA();
+      }
+      cfg.runCopyPropagation();
+    });
+
     test('Query livein', () {
       if (!cfg.hasPhiNodes) {
         cfg.insertPhiNodes();
@@ -203,6 +213,70 @@ return x₁\n
       expect(cfg.isLiveOut(x2, block2), false);
     });
 
+    test('Remove unused defines', () {
+      if (!cfg.hasPhiNodes) {
+        cfg.insertPhiNodes();
+      }
+      if (!cfg.inSSAForm) {
+        cfg.computeSemiPrunedSSA();
+      }
+      cfg.removeUnusedDefines();
+      expect(() => {print(cfg)}, prints('''
+B0:
+x₀ = imm 0  
+i₀ = imm 0
+→ (B1)
+
+B1:
+i₁ = φ(i₀, i₂)  
+x₁ = φ(x₀, x₂)  
+n = imm 10  
+@branch = i₁ < n
+→ (B2, B3)
+
+B2:
+x₂ = x₁ + i₁  
+@1 = imm 1  
+i₂ = i₁ + @1
+→ (B1)
+
+B3:
+return x₁\n
+'''));
+    });
+
+    test('Remove empty and unused blocks', () {
+      if (!cfg.hasPhiNodes) {
+        cfg.insertPhiNodes();
+      }
+      if (!cfg.inSSAForm) {
+        cfg.computeSemiPrunedSSA();
+      }
+      cfg.removeEmptyAndUnusedBlocks();
+      expect(() => {print(cfg)}, prints('''
+B0:
+x₀ = imm 0  
+i₀ = imm 0
+→ (B1)
+
+B1:
+i₁ = φ(i₀, i₂)  
+x₁ = φ(x₀, x₂)  
+n = imm 10  
+@branch = i₁ < n
+→ (B2, B3)
+
+B2:
+x₂ = x₁ + i₁  
+@1 = imm 1  
+i₂ = i₁ + @1
+→ (B1)
+
+B3:
+return x₁\n
+'''));
+    });
+
     test('Remove phi nodes', () {
       if (!cfg.hasPhiNodes) {
         cfg.insertPhiNodes();
@@ -211,12 +285,11 @@ return x₁\n
         cfg.computeSemiPrunedSSA();
       }
       cfg.removePhiNodes((l, r) => Assign(l, r));
+      print(cfg);
       expect(() => print(cfg), prints('''
 B0:
-x₀ = imm 0  
-i₀ = imm 0  
-i₁ = i₀  
-x₁ = x₀
+x₁ = imm 0  
+i₁ = imm 0
 → (B1)
 
 B1:
@@ -225,11 +298,9 @@ n = imm 10
 → (B2, B3)
 
 B2:
-x₂ = x₁ + i₁  
+x₁ = x₁ + i₁  
 @1 = imm 1  
-i₂ = i₁ + @1  
-i₁ = i₂  
-x₁ = x₂
+i₁ = i₁ + @1
 → (B1)
 
 B3:
@@ -382,6 +453,224 @@ return x₁\n
         cfg.insertPhiNodes();
       }
       cfg.computeSemiPrunedSSA();
+      expect(() => {print(cfg)}, prints('''
+c1(0):
+a₀ = imm 0  
+b₀ = imm 0
+→ (c2(1))
+
+c2(1):
+a₁ = φ(a₀, a₂)  
+b₁ = φ(b₀, b₁₀)  
+a₂ = imm 2
+→ (c3(2), c11(10))
+
+c3(2):
+b₂ = imm 3  
+@branch = a₂ < b₂
+→ (c4(3), c8(4))
+
+c11(10):
+
+c4(3):
+b₃ = a₂
+→ (c5(5))
+
+c8(4):
+b₄ = φ(b₂, b₈)  
+b₅ = imm 20
+→ (c9(8))
+
+c5(5):
+b₆ = φ(b₃, b₉)  
+b₇ = imm 10
+→ (c6(6))
+
+c9(8):
+i₀ = imm 0  
+b₈ = i₀
+→ (c6(6), c10(9))
+
+c6(6):
+b₉ = φ(b₇, b₈)  
+i₁ = φ(i₀)  
+@branch = b₉ < a₂
+→ (c5(5), c7(7))
+
+c10(9):
+@branch = b₈ < i₀
+→ (c8(4))
+
+c7(7):
+i₂ = imm 0  
+b₁₀ = i₂
+→ (c2(1))\n\n
+'''));
+    });
+
+    bool ranCopyPropagation = false,
+        removedUnusedDefines = false,
+        removedBlocks = false;
+
+    test('Run copy propagation', () {
+      if (!cfg.hasPhiNodes) {
+        cfg.insertPhiNodes();
+      }
+      if (!cfg.inSSAForm) {
+        cfg.computeSemiPrunedSSA();
+      }
+      cfg.runCopyPropagation();
+      ranCopyPropagation = true;
+      expect(() => {print(cfg)}, prints('''
+c1(0):
+a₀ = imm 0  
+b₀ = imm 0
+→ (c2(1))
+
+c2(1):
+a₁ = φ(a₀, a₂)  
+b₁ = φ(b₀, i₂)  
+a₂ = imm 2
+→ (c3(2))
+
+c3(2):
+b₂ = imm 3  
+@branch = a₂ < b₂
+→ (c4(3), c8(4))
+
+c4(3):
+b₃ = a₂
+→ (c5(5))
+
+c8(4):
+b₄ = φ(b₂, i₀)  
+b₅ = imm 20
+→ (c9(8))
+
+c5(5):
+b₆ = φ(b₉, a₂)  
+b₇ = imm 10
+→ (c6(6))
+
+c9(8):
+i₀ = imm 0  
+b₈ = i₀
+→ (c6(6))
+
+c6(6):
+b₉ = φ(b₇, i₀)  
+i₁ = φ(i₀)  
+@branch = b₉ < a₂
+→ (c5(5), c7(7))
+
+c7(7):
+i₂ = imm 0  
+b₁₀ = i₂
+→ (c2(1))\n\n
+'''));
+    });
+
+    test('Remove unused defines', () {
+      if (!cfg.hasPhiNodes) {
+        cfg.insertPhiNodes();
+      }
+      if (!cfg.inSSAForm) {
+        cfg.computeSemiPrunedSSA();
+      }
+      if (!ranCopyPropagation) {
+        cfg.runCopyPropagation();
+      }
+      cfg.removeUnusedDefines();
+      removedUnusedDefines = true;
+      expect(() => {print(cfg)}, prints('''
+c1(0):
+a₀ = imm 0  
+b₀ = imm 0
+→ (c2(1))
+
+c2(1):
+a₂ = imm 2
+→ (c3(2))
+
+c3(2):
+b₂ = imm 3  
+@branch = a₂ < b₂
+→ (c4(3), c8(4))
+
+c4(3):
+
+→ (c5(5))
+
+c8(4):
+
+→ (c9(8))
+
+c5(5):
+b₇ = imm 10
+→ (c6(6))
+
+c9(8):
+i₀ = imm 0
+→ (c6(6))
+
+c6(6):
+b₉ = φ(b₇, i₀)  
+@branch = b₉ < a₂
+→ (c5(5), c7(7))
+
+c7(7):
+i₂ = imm 0
+→ (c2(1))\n\n
+'''));
+    });
+
+    test('Remove empty and unused blocks', () {
+      if (!cfg.hasPhiNodes) {
+        cfg.insertPhiNodes();
+      }
+      if (!cfg.inSSAForm) {
+        cfg.computeSemiPrunedSSA();
+      }
+      if (!ranCopyPropagation) {
+        cfg.runCopyPropagation();
+      }
+      if (!removedUnusedDefines) {
+        cfg.removeUnusedDefines();
+      }
+      cfg.removeEmptyAndUnusedBlocks();
+      removedBlocks = true;
+      expect(() => {print(cfg)}, prints('''
+c1(0):
+a₀ = imm 0  
+b₀ = imm 0
+→ (c2(1))
+
+c2(1):
+a₂ = imm 2
+→ (c3(2))
+
+c3(2):
+b₂ = imm 3  
+@branch = a₂ < b₂
+→ (c5(5), c9(8))
+
+c5(5):
+b₇ = imm 10
+→ (c6(6))
+
+c9(8):
+i₀ = imm 0
+→ (c6(6))
+
+c6(6):
+b₉ = φ(b₇, i₀)  
+@branch = b₉ < a₂
+→ (c5(5), c7(7))
+
+c7(7):
+i₂ = imm 0
+→ (c2(1))\n\n
+'''));
     });
 
     test('Remove phi nodes', () {
@@ -391,7 +680,47 @@ return x₁\n
       if (!cfg.inSSAForm) {
         cfg.computeSemiPrunedSSA();
       }
+      if (!ranCopyPropagation) {
+        cfg.runCopyPropagation();
+      }
+      if (!removedUnusedDefines) {
+        cfg.removeUnusedDefines();
+      }
+      if (!removedBlocks) {
+        cfg.removeEmptyAndUnusedBlocks();
+      }
       cfg.removePhiNodes((l, r) => Assign(l, r));
+      expect(() => print(cfg), prints('''
+c1(0):
+a₀ = imm 0  
+b₀ = imm 0
+→ (c2(1))
+
+c2(1):
+a₂ = imm 2
+→ (c3(2))
+
+c3(2):
+b₂ = imm 3  
+@branch = a₂ < b₂
+→ (c5(5), c9(8))
+
+c5(5):
+b₉ = imm 10
+→ (c6(6))
+
+c9(8):
+b₉ = imm 0
+→ (c6(6))
+
+c6(6):
+@branch = b₉ < a₂
+→ (c5(5), c7(7))
+
+c7(7):
+i₂ = imm 0
+→ (c2(1))\n\n
+'''));
     });
   });
 }