+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+- Improved SSA renaming algorithm.
+- Added liveness computations
+- Added copy propagation, unused defines elimination, and dead block 
+  elimination optimizations.
+
 ## 1.0.0
 
 - Initial version.